Added:AesZipDecrypt.

+++ src/com/duole/utils/JsonUtils.java
@@ -5,12 +5,12 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
 
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import android.util.Log;
 
-import com.duole.Duole;
 import com.duole.pojos.asset.Asset;
 
 public class JsonUtils {
@@ -18,28 +18,34 @@ public class JsonUtils {
 	public static void parserJson(ArrayList<Asset> alAsset ,JSONObject jsonObject) throws MalformedURLException, JSONException{
 		File client;
 		String version = jsonObject.getString("ver");
-		client = new File(Constants.CacheDir + "client.apk");
-//		Log.v("TAG", "server version" + version);
-//		Log.v("TAG", "local version" + Constants.System_ver);
-		if(!version.equals(Constants.System_ver)){
+		if(!version.equals("null") && !version.equals("")){
 			client = new File(Constants.CacheDir + "client.apk");
-//			Log.v("TAG", "apk version " + DuoleUtils.getPackageVersion(client));
-			if(client.exists()){
-				if(!version.equals(DuoleUtils.getPackageVersion(client)) && !Constants.clientApkDownloaded){
+			if(!version.equals(Constants.System_ver)){
+				client = new File(Constants.CacheDir + "client.apk");
+				if(client.exists()){
+					if(!version.equals(DuoleUtils.getPackageVersion(client)) && !Constants.clientApkDownloaded){
+						DuoleUtils.updateClient();
+					}
+				}else{
 					DuoleUtils.updateClient();
 				}
 			}else{
-				DuoleUtils.updateClient();
+				client = new File(Constants.CacheDir + "client.apk");
+				if(client.exists()){
+					client.delete();
+				}
 			}
-		}else{
-			client = new File(Constants.CacheDir + "client.apk");
-			if(client.exists()){
-				client.delete();
-			}
+		}
+		
+		if(jsonObject.has("front")){
+			JSONArray jsonArray = jsonObject.getJSONArray("front");
+			
+			Log.v("TAG", jsonArray.length() + " front length");
 		}
 		for (int i = 0;i< jsonObject.length(); i++) {
 			try{
 				JSONObject jsonItem = jsonObject.getJSONObject("item" + i);
+				
 				if (jsonItem != null) {
 					Asset asset = new Asset(jsonItem);
 					alAsset.add(asset);
@@ -47,9 +53,29 @@ public class JsonUtils {
 			}catch(Exception e){
 				break;
 			}
-			
+			try {
+				JSONObject jsonFront = jsonObject.getJSONObject("front" + i);
+				if (jsonFront != null) {
+					Asset asset = new Asset(jsonFront);
+					alAsset.add(asset);
+				}
+			} catch (Exception e) {
+			}
 			
 		}
+		
+		try{
+			if(jsonObject.has("front")){
+				JSONArray jsonArray = jsonObject.getJSONArray("front");
+				
+				Log.v("TAG", jsonArray.length() + " front length");
+			}
+			
+		}catch(Exception e){
+			e.printStackTrace();
+		}
+		
+		
 		File file = null;
 		if(!Constants.bgurl.equals("")){
 			file = new File(Constants.CacheDir
